channels/matrix/pyproject.toml
[project]
name = "letsgo-channel-matrix"
version = "0.1.0"
description = "Matrix channel adapter for LetsGo gateway"
requires-python = ">=3.11"
dependencies = [
    "letsgo-gateway",
]

[project.optional-dependencies]
nio = ["matrix-nio>=0.21"]

[project.entry-points."letsgo.channels"]
matrix = "letsgo_channel_matrix:MatrixChannel"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["letsgo_channel_matrix"]
